HPCC-8159 Prepare for use of precompiled headers in eclcc-generated code
Reorganise the generated code to make it easier to use precompiled headers.
It also has the benefit of making the generated code a little easier to
read.

Modifying jcomp to actually use precompiled headers will follow in a second
commit.
